pyproject.toml

[project]
name = "octoagentteam"
version = "0.1.0"
description = "OCTOAgentTeam – ADK multi‑agent reference"
requires-python = ">=3.10"
dependencies = [
  "google-adk>=0.0.5",
  "pytest",
  "google-generativeai",
]

[build-system]
requires = ["setuptools","wheel"]
build-backend = "setuptools.build_meta"
